refactor: consolidate settings management into ShowSettings class and remove SettingsStore
